Use Arc<RwLock<>> for router state
This means the state doesn't have to be cloned (even though the cloning
would be minimal since obviously the DB connection pool is just a
reference to the DB). I think we still should use SqlitePool instead
of a connection struct since from what I can see this is the more
idiomatic method.